feat: refactor permissions service to features service
