+++ packages/snjs/lib/storage_keys.ts
@@ -37,7 +37,6 @@ export enum StorageKey {
   StorageEncryptionPolicy = 'storage_policy',
   WebSocketUrl = 'webSocket_url',
   UserRoles = 'user_roles',
-  UserPermissions = 'user_permissions'
 }
 
 export enum NonwrappedStorageKey {
